refactor: EmailForm에서 인증 메일 요청을 보내도록 변경
